+++ VoxelEngine/VoxelEngine/VoxelEngine/Graphics/stbiImpl.cpp
@@ -1,0 +1,2 @@
+#define STB_IMAGE_IMPLEMENTATION
+#include <stb_image.h>
