+++ VoxelEngine/VoxelEngine/VoxelEngine/Headers/Components/AllComponentHeaders.h
@@ -1,5 +1,5 @@
 #ifndef COMPONENT_COUNT
-#define COMPONENT_COUNT 5
+#define COMPONENT_COUNT 4
 #endif
 
 #ifdef FACTORY_RUNNING
@@ -7,7 +7,6 @@
 #include "../Headers/Factory.h"
 #include "Camera.h"
 #include "StubComponent.h"
-#include "Tag.h"
 #include "TestingComponent.h"
 #include "VoxelWorld.h"
 
@@ -18,8 +17,6 @@ Factory::ComponentPropertyMap["Camera"] = std::vector<PropertyID>({
   });
 Factory::ComponentPropertyMap["StubComponent"] = std::vector<PropertyID>({
   });
-Factory::ComponentPropertyMap["Tag"] = std::vector<PropertyID>({
-  });
 Factory::ComponentPropertyMap["TestingComponent"] = std::vector<PropertyID>({
   PropertyID("componentData", TestingComponent::componentData_id, offsetof(TestingComponent, componentData), sizeof(Int)),
   });
@@ -29,9 +26,8 @@ Factory::ComponentPropertyMap["VoxelWorld"] = std::vector<PropertyID>({
 
 inline std::unique_ptr<Component> Component0() { return std::move(Camera::RegisterCamera()); }
 inline std::unique_ptr<Component> Component1() { return std::move(StubComponent::RegisterStubComponent()); }
-inline std::unique_ptr<Component> Component2() { return std::move(Tag::RegisterTag()); }
-inline std::unique_ptr<Component> Component3() { return std::move(TestingComponent::RegisterTestingComponent()); }
-inline std::unique_ptr<Component> Component4() { return std::move(VoxelWorld::RegisterVoxelWorld()); }
+inline std::unique_ptr<Component> Component2() { return std::move(TestingComponent::RegisterTestingComponent()); }
+inline std::unique_ptr<Component> Component3() { return std::move(VoxelWorld::RegisterVoxelWorld()); }
 
 
 #endif // !FACTORY_RUNNING
