+++ VoxelEngine/VoxelEngine/VoxelEngine/Voxels/Refactor/sun.cpp
@@ -7,7 +7,6 @@
 #include <Refactor/settings.h>
 #include <limits>
 #include <Graphics/Vertices.h>
-#include <Rendering/Renderer.h>
 
 #include <Systems/GraphicsSystem.h>
 #include <Graphics/GraphicsIncludes.h>
@@ -41,8 +40,6 @@ void Sun::Update()
 	
 	//glm::mat4 lightView = glm::lookAt(pos_, dir_, glm::vec3(0.0, 1.0, 0.0));
 	view_ = glm::lookAt(pos_, glm::normalize(GetCurrentCamera()->GetDir()), glm::vec3(0.0, 1.0, 0.0));
-
-	dirLight_.Update(pos_, dir_);
 }
 
 //void Sun::Init()
