+++ VoxelEngine/VoxelEngine/VoxelEngine/Voxels/Rendering/NuRenderer.h
@@ -1,6 +1,9 @@
 #pragma once
+#include <Refactor/sun.h>
+#include <memory>
 
 class TextureArray;
+class Sun;
 
 namespace NuRenderer
 {
@@ -15,7 +18,13 @@ namespace NuRenderer
 	void splatChunks();
 	void drawChunksWater();
 
+	// generic drawing functions
+	void drawAxisIndicators();
+	void drawQuad();
+	void DrawCube();
+
 	inline int drawCalls = 0;
+	inline std::unique_ptr<Sun> activeSun_;
 
 	struct Settings
 	{
